feat: refactor frequency_map and huffman_tree build
