+++ src/encoding_map/mod.rs
@@ -1,0 +1,5 @@
+pub mod frequency_map;
+pub mod huffman_tree;
+
+#[cfg(test)]
+mod tests;
