+++ src/encoding_map/tests/mod.rs
@@ -1,0 +1,2 @@
+mod frequency_map;
+mod huffman_tree;
